Made changes to have sign in contained in component
