+++ frontend/src/components/Chatterbox/LoginForm/loginform.jsx
@@ -28,7 +28,7 @@ const LoginForm = () => {
 	return (
 		<div className="wrapper">
 			<div className="form">
-				<h1 className="title">ChatterBot</h1>
+				<h1 className="title">Chatterbox</h1>
 				<form onSubmit={handleSubmit}>
 					<input type="text" value={username} onChange={(e) => setUsername(e.target.value)} className="input" placeholder="Username" required />
 					<input type="password" value={password} onChange={(e) => setPassword(e.target.value)} className="input" placeholder="Password" required />
